refactor(test): P5-1 follow-up — extract RN style-prop bridge primitives
Continue test_widget_bridge.cpp split. Extracts the 626-line pulp
#1026 RN style-prop cluster into a sibling TU.

Coverage: the RN-shaped setShadow / setOpacity / setTransform style
functions that flow through bridge into View's geometry / shadow /
transform / opacity slots. Includes RN's documented quirks like
shadowOpacity composed into the alpha channel of shadowColor, and
the transform-prop aggregation that turns individual scale / rotate /
translate calls into a single uniform setTransform.

* test_widget_bridge.cpp: 7,475 → 6,849 lines (-626)
* New target pulp-test-widget-bridge-rn-style: 666 lines.

This is commit 1 of 2 in a bundled-cuts PR (Tier-4 OOS pins follow
in commit 2; single version bump at the end).
